pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.11.0.post1",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "openai>=1.99.3",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "google-genai>=1.29.0",
]
